conda-requirements.txt
newspaper3k=0.2.8
requests=2.25.1
textblob=0.15.3
tqdm=4.56.0
pydub=0.23.1
gunicorn=20.0.4
ffmpeg=4.3.1
redis-py=3.5.3
aiohttp=3.7.3
python=3.8.6
sanic=20.12.1
